Some basic tests of list item functionality
This commit contains a function for creating a simple list of items.
The purpose was to test out a couple of things about rendering
documents within a test setting. This enables us to test out
basic creation functions that we are going to use for creating the
bulk of the store.