Buggy literal dom version
This is a "final commit" for chapter 4 of grokking simplicity.
Basically we've bootstrapped the project using vanilly js.
Just wanted to see a very rudimentary version of our app running in the browser.
Basically we're starting to run in to the common problems of code that is not
very orthogonal - we've got logic mixed with state and ui stuff. A framework
like react is really where we want to go at this point (and probably was to
begin with actually). From here we'll create another branch to handle the
migration of the existing code and resources and merge it back in to the
dev and main branches when we're done. Over and out.